Move utils and setting_defaults to helpers
