+++ app/javascript/react_app/components/group/search_bar.jsx
@@ -4,7 +4,7 @@ import { bindActionCreators } from 'redux';
 import { Link } from 'react-router-dom';
 
 import { fetchSearchBirds, clearSearchBirds } from '../../actions';
-import { hashify } from '../../utils';
+import { hashify } from '../../helpers/utils';
 
 class SearchBar extends Component {
   constructor(props) {
