+++ app/javascript/react_app/helpers/setting_defaults.js
