+++ app/javascript/react_app/helpers/utils.jsx
